refactor: memeber refactor
